+++ examples/core/main.tf
@@ -10,7 +10,7 @@ module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
   version = "~> 2.44"
 
-  name = "simple-vpc"
+  name = "simple-waf-test-vpc"
 
   cidr = "10.0.0.0/16"
 
@@ -35,69 +35,6 @@ module "alb" {
   subnets            = module.vpc.public_subnets
 }
 
-resource "aws_lb_listener" "alb_80" {
-  load_balancer_arn = module.alb.arn
-  port              = "80"
-  protocol          = "HTTP"
-
-  default_action {
-    type             = "forward"
-    target_group_arn = module.fargate.target_group_arn
-  }
-}
-
-#####
-# Security Group Config
-#####
-resource "aws_security_group_rule" "alb_ingress_80" {
-  security_group_id = module.alb.security_group_id
-  type              = "ingress"
-  protocol          = "tcp"
-  from_port         = 80
-  to_port           = 80
-  cidr_blocks       = ["0.0.0.0/0"]
-  ipv6_cidr_blocks  = ["::/0"]
-}
-
-resource "aws_security_group_rule" "task_ingress_80" {
-  security_group_id        = module.fargate.service_sg_id
-  type                     = "ingress"
-  protocol                 = "tcp"
-  from_port                = 80
-  to_port                  = 80
-  source_security_group_id = module.alb.security_group_id
-}
-
-#####
-# ECS cluster and fargate
-#####
-resource "aws_ecs_cluster" "cluster" {
-  name = "example-ecs-cluster"
-}
-
-module "fargate" {
-  source  = "umotif-public/ecs-fargate/aws"
-  version = "~> 3.0.0"
-
-  name_prefix        = "ecs-fargate-example"
-  vpc_id             = module.vpc.vpc_id
-  private_subnet_ids = module.vpc.public_subnets
-  lb_arn             = module.alb.arn
-  cluster_id         = aws_ecs_cluster.cluster.id
-
-  task_container_image   = "marcincuber/2048-game:latest"
-  task_definition_cpu    = 256
-  task_definition_memory = 512
-
-  task_container_port             = 80
-  task_container_assign_public_ip = true
-
-  health_check = {
-    port = "traffic-port"
-    path = "/"
-  }
-}
-
 #####
 # Web Application Firewall configuration
 #####
@@ -106,6 +43,8 @@ module "waf" {
 
   name_prefix = "test-waf-setup"
   alb_arn     = module.alb.arn
+
+  allow_default_action = true
 
   create_alb_association = true
 
@@ -119,6 +58,8 @@ module "waf" {
     {
       name     = "AWSManagedRulesCommonRuleSet-rule-1"
       priority = "1"
+
+      override_action = "none"
 
       visibility_config = {
         cloudwatch_metrics_enabled = false
@@ -139,6 +80,8 @@ module "waf" {
     {
       name     = "AWSManagedRulesKnownBadInputsRuleSet-rule-2"
       priority = "2"
+
+      override_action = "count"
 
       visibility_config = {
         cloudwatch_metrics_enabled = false